Attempt to keep viewport top when zooming font
A changed font size causes the window position to be wrong since the
window is positioned in pixels.  When the font changes the window needs
to be scrolled to keep the same line number at the top of the screen,
arguably.  But maybe it should try to keep the cursor in the same relative
position on the screen.  Or the center line on the screen. Or the bottom
of the screen.  It's sort of arbitrary.

The current code may cause the buffer to "scroll" by several hundred lines
or more, which is very disconcerting.  If the top row of the buffer remains
consistent at least the user can tell what is happening and can adjust
the buffer to his liking more easily after he is done.

But this still doesn't handle the invisible buffers, so it's not really
a solution yet.